load caches from checkpoint
